Fix Windows build error: align io_expander read_*_pins to uint16_t

+++ components/bsp/include/bsp_board.h
@@ -196,8 +196,8 @@ typedef struct {
     esp_err_t (*init)(uint8_t i2c_addr);
     esp_err_t (*set_direction)(uint8_t pin, bool is_output);
     esp_err_t (*set_level)(uint8_t pin, bool level);
-    esp_err_t (*read_output_pins)(uint8_t *pin_val);
-    esp_err_t (*read_input_pins)(uint8_t *pin_val);
+    esp_err_t (*read_output_pins)(uint16_t *pin_val);
+    esp_err_t (*read_input_pins)(uint16_t *pin_val);
     esp_err_t (*multi_write_start)(void);
     esp_err_t (*multi_write_new_level)(int pin, bool new_level);
     esp_err_t (*multi_write_end)(void);
